Refact: Перенес тип проектов в папку модели
