+++ src/model/data-types.ts
@@ -1,0 +1,8 @@
+export interface Project {
+  id: string
+  userId: string
+  title: string
+  description: string
+  time: Date
+  newComments: number
+}
